add type folder and change import paths
