+++ src/type/Category.ts
@@ -1,0 +1,4 @@
+export type CategoryType = {
+  id: number;
+  category: string;
+};
