+++ src/type/Todo.ts
@@ -1,0 +1,10 @@
+export type TodoType = {
+  id: number;
+  title: string;
+  description: string;
+  deadline: Date;
+  categories: {
+    id: number;
+    category: string;
+  }[];
+};
